+++ include/Menu.h
@@ -1,6 +1,6 @@
 /*!
  * \file include/Menu.h
- * \brief Menu 'overlay' interface
+ * \brief Main Menu
  *
  * \author xythobuz
  */
@@ -8,57 +8,30 @@
 #ifndef _MENU_H_
 #define _MENU_H_
 
-#include <functional>
+#include <vector>
+
+#include "Script.h"
+#include "utils/Folder.h"
 
 class Menu {
   public:
+    static int initialize();
+    static void shutdown();
+    static void display();
 
-    virtual ~Menu() { }
-
-    virtual int initialize() = 0;
-
-    virtual void display() = 0;
-
-    virtual void handleKeyboard(KeyboardButton key, bool pressed) = 0;
-
-    virtual void handleMouseClick(unsigned int x, unsigned int y,
-                                  KeyboardButton button, bool released) = 0;
-
-    virtual void handleMouseScroll(int xrel, int yrel) = 0;
-
-    bool isVisible() { return visible; }
-
-    void setVisible(bool v) { visible = v; }
+    static bool isVisible() { return visible; }
+    static void setVisible(bool v) { visible = v; }
 
     static const glm::vec4 textColor;
     static const glm::vec4 selectedColor;
 
-  protected:
-
-    virtual void showDialog(std::string msg, std::string btn1, std::string btn2 = "",
-                            std::function<int (bool state)> callback = std::function<int (bool)>());
-
-    virtual void ackDialog();
-
-    virtual bool handleKeyboardDialog(KeyboardButton key, bool pressed);
-
-    virtual bool handleMouseClickDialog(unsigned int x, unsigned int y,
-                                        KeyboardButton button, bool released);
-
-    virtual bool handleMouseScrollDialog(int xrel, int yrel);
-
-    virtual void displayDialog();
-
-    bool dialogState;
-    std::string dialogText;
-    std::string dialogButton1;
-    std::string dialogButton2;
-    std::function<int (bool state)> dialogFunction;
-
-    bool visible;
+  private:
+    static bool visible;
+    static Folder* mapFolder;
+    static std::vector<Script> scripts;
+    static std::vector<Folder> paths;
+    static std::vector<int> images;
 };
-
-Menu& getMenu();
 
 #endif