+++ include/utils/strings.h
@@ -16,5 +16,11 @@ std::string expandHomeDirectory(std::string s);
 
 bool stringEndsWith(std::string s, std::string suffix, bool casesensitive = false);
 
+std::string removeLastPathElement(std::string s);
+
+std::string getLastPathElement(std::string s);
+
+std::string convertPathDelimiter(std::string s);
+
 #endif