+++ src/Menu.cpp
@@ -1,173 +1,230 @@
 /*!
  * \file src/Menu.cpp
- * \brief Menu 'overlay'
+ * \brief Main Menu
  *
  * \author xythobuz
  */
 
+#include "imgui/imgui.h"
+
 #include "global.h"
+#include "Console.h"
+#include "Game.h"
 #include "Log.h"
-#include "Menu.h"
+#include "RunTime.h"
+#include "UI.h"
+#include "TextureManager.h"
 #include "system/Font.h"
 #include "system/Window.h"
+#include "utils/random.h"
+#include "utils/strings.h"
+#include "Menu.h"
 
 const glm::vec4 Menu::textColor(0.5f, 0.7f, 1.0f, 1.0f);
 const glm::vec4 Menu::selectedColor(1.0f, 0.0f, 0.0f, 1.0f);
 
-void Menu::showDialog(std::string msg, std::string btn1, std::string btn2,
-                      std::function<int (bool state)> callback) {
-    // Only show one dialog at a time
-    assertEqual(dialogText.length(), 0);
-    assertEqual(dialogButton1.length(), 0);
-    assertEqual(dialogButton2.length(), 0);
+bool Menu::visible = false;
+Folder* Menu::mapFolder = nullptr;
+std::vector<Script> Menu::scripts;
+std::vector<Folder> Menu::paths;
+std::vector<int> Menu::images;
 
-    assertGreaterThan(msg.length(), 0);
-    assertGreaterThan(btn1.length(), 0);
+int Menu::initialize() {
+    shutdown();
+    mapFolder = new Folder(RunTime::getPakDir());
 
-    dialogText = msg;
-    dialogButton1 = btn1;
-    dialogButton2 = btn2;
-    dialogState = false;
-    dialogFunction = callback;
+    std::vector<File> files;
+    mapFolder->findRecursiveFilesEndingWith(files, "tombpc.dat");
+    for (auto& f : files) {
+        scripts.emplace_back();
+        auto s = scripts.size() - 1;
 
-    Log::get(LOG_USER) << dialogText << Log::endl;
-}
+        images.push_back(-1);
+        auto i = images.size() - 1;
 
-void Menu::ackDialog() {
-    dialogText = "";
-    dialogButton1 = "";
-    dialogButton2 = "";
+        std::string path = convertPathDelimiter(removeLastPathElement(f.getPath()));
+        paths.emplace_back(path);
+        auto p = paths.size() - 1;
 
-    if (dialogFunction) {
-        int error = dialogFunction(dialogState);
-        if (error != 0) {
-            showDialog("Error processing dialog callback!", "OK", "");
+        if (scripts.at(s).load(f.getPath()) != 0) {
+            Log::get(LOG_ERROR) << "Invalid Script: \"" << f.getPath()
+                                << "\"!" << Log::endl;
+            scripts.pop_back();
+        } else {
+            // Load one of the pictures of this level
+            std::vector<File> texFiles;
+            paths.at(p).findRecursiveFilesEndingWith(texFiles, ".pcx");
+            paths.at(p).findRecursiveFilesEndingWith(texFiles, ".bmp");
+            paths.at(p).findRecursiveFilesEndingWith(texFiles, ".png");
+            paths.at(p).findRecursiveFilesEndingWith(texFiles, ".tga");
+            paths.at(p).findRecursiveFilesEndingWith(texFiles, ".jpg");
+
+            if (texFiles.size() > 0) {
+                int which = randomInteger(texFiles.size() - 1);
+                images.at(i) = TextureManager::loadImage(texFiles.at(which).getPath(),
+                               TextureStorage::SYSTEM);
+            }
         }
     }
 
-    dialogState = false;
+    return 0;
 }
 
-bool Menu::handleKeyboardDialog(KeyboardButton key, bool pressed) {
-    if (!pressed)
-        return (dialogText.length() > 0);
-
-    if (dialogText.length() > 0) {
-        if (dialogButton2.length() == 0) {
-            if (key == enterKey) {
-                ackDialog();
-            }
-        } else {
-            if (key == enterKey) {
-                ackDialog();
-            } else if (key == leftKey) {
-                dialogState = !dialogState;
-            } else if (key == rightKey) {
-                dialogState = !dialogState;
-            } else if (key == upKey) {
-                dialogState = !dialogState;
-            } else if (key == downKey) {
-                dialogState = !dialogState;
-            }
-        }
-        return true;
+void Menu::shutdown() {
+    if (mapFolder != nullptr) {
+        delete mapFolder;
+        mapFolder = nullptr;
     }
 
-    return false;
+    scripts.clear();
+    paths.clear();
+    images.clear(); //! \TODO free texture slots
 }
 
-bool Menu::handleMouseClickDialog(unsigned int x, unsigned int y,
-                                  KeyboardButton button, bool released) {
-    //! \todo Allow mouse usage of Menu dialogs!
-    return (dialogText.length() > 0);
-}
+void Menu::display() {
+    if (!visible)
+        return;
 
-bool Menu::handleMouseScrollDialog(int xrel, int yrel) {
-    //! \todo Allow mouse usage of Menu dialogs!
-    return (dialogText.length() > 0);
-}
+    glm::vec2 size = Window::getSize();
+    if (!ImGui::Begin("Menu", nullptr, ImVec2(size.x, size.y), -1.0f, ImGuiWindowFlags_NoTitleBar
+                      | ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoCollapse
+                      | ImGuiWindowFlags_NoSavedSettings)) {
+        ImGui::End();
+        return;
+    }
 
-void Menu::displayDialog() {
-    if (dialogText.length() > 0) {
-        unsigned int wMax = ((unsigned int)(Window::getSize().x * 0.66f));
+    ImGui::SetWindowPos(ImVec2(0, 0));
+    ImGui::SetWindowSize(ImVec2(size.x, size.y));
 
-        unsigned int w0 = Font::widthText(1.0f, dialogText) + 20;
-        if (w0 > wMax)
-            w0 = wMax;
-        unsigned int h0 =  Font::heightText(1.0f, w0, dialogText) + 10;
+    // Draw heading with version string
+    auto headingSize = ImGui::CalcTextSize(VERSION);
+    headingSize.x *= 3.0f; headingSize.y *= 3.0f;
+    ImGui::SetWindowFontScale(3.0f);
+    ImGui::SameLine(0, (size.x / 2) - (headingSize.x / 2));
+    ImGui::TextColored(textColor, VERSION);
+    ImGui::SetWindowFontScale(1.0f);
 
-        assertGreaterThan(dialogButton1.length(), 0);
-        unsigned int w1 = Font::widthText(1.0f, dialogButton1) + 20;
-        if (w1 > wMax)
-            w1 = wMax;
-        unsigned int h1 = Font::heightText(1.0f, w1, dialogButton1) + 10;
-
-        unsigned int wOverlay = wMax, hOverlay, w2 = 0;
-
-        if (dialogButton2.length() > 0) {
-            // Show text and two buttons
-            w2 = Font::widthText(1.0f, dialogButton2) + 20;
-            if (w2 > wMax)
-                w2 = wMax;
-            unsigned int h2 = Font::heightText(1.0f, w2, dialogButton2) + 10;
-
-            if (w0 > (w1 + w2)) {
-                if (w0 < wMax) {
-                    wOverlay = w0;
-                }
-            } else if (w0 < (w1 + w2)) {
-                if ((w1 + w2) < wMax) {
-                    wOverlay = (w1 + w2);
-                }
-            }
-
-            if ((w1 + w2) <= wMax) {
-                hOverlay = h0 + ((h1 + h2) / 2);
-            } else {
-                hOverlay = h0 + h1 + h2;
-            }
-        } else {
-            // Show text and one button
-            if (w0 > w1) {
-                if (w0 < wMax) {
-                    wOverlay = w0;
-                }
-            } else if (w0 < w1) {
-                if (w1 < wMax) {
-                    wOverlay = w1;
-                }
-            }
-
-            hOverlay = h0 + h1;
-        }
-
-        unsigned int xOverlay = (Window::getSize().x - wOverlay) / 2;
-        unsigned int yOverlay = (Window::getSize().y - hOverlay) / 2;
-
-        /*
-        glColor4f(0.0f, 0.0f, 0.0f, 0.75f);
-        glDisable(GL_TEXTURE_2D);
-        glRecti(xOverlay, yOverlay, xOverlay + wOverlay, yOverlay + hOverlay);
-        glEnable(GL_TEXTURE_2D);
-        */
-
-        Font::drawTextWrapped(xOverlay + 10, yOverlay + 5, 1.0f, textColor, w0, dialogText);
-        if (dialogButton2.length() > 0) {
-            if ((w1 + w2) <= wMax) {
-                Font::drawTextWrapped(xOverlay + 10, yOverlay + 10 + h0, 1.0f,
-                                      dialogState ? textColor : selectedColor, w1, dialogButton1);
-                Font::drawTextWrapped(xOverlay + 10 + w1, yOverlay + 10 + h0,
-                                      1.0f, dialogState ? selectedColor : textColor, w2, dialogButton2);
-            } else {
-                Font::drawTextWrapped((Window::getSize().x - w1) / 2,
-                                      yOverlay + 10 + h0, 1.0f, dialogState ? textColor : selectedColor, w1, dialogButton1);
-                Font::drawTextWrapped((Window::getSize().x - w2) / 2,
-                                      yOverlay + 10 + h0 + h1, 1.0f, dialogState ? selectedColor : textColor, w2, dialogButton2);
-            }
-        } else {
-            Font::drawTextWrapped((Window::getSize().x - w1) / 2,
-                                  yOverlay + 10 + h0, 1.0f, selectedColor, w1, dialogButton1);
+    static ImGuiFs::Dialog* dialog = nullptr;
+    const bool browseButtonPressed = ImGui::Button("Select Level File");
+    if (browseButtonPressed) {
+        if (dialog != nullptr)
+            delete dialog;
+        dialog = new ImGuiFs::Dialog();
+    }
+    if (dialog) {
+        dialog->chooseFileDialog(browseButtonPressed, RunTime::getPakDir().c_str(),
+                                 ".phd;.psx;.tub;.tr2", nullptr, ImVec2(400, 300), ImGui::GetMousePos());
+        std::string selectedFile = dialog->getChosenPath();
+        if (selectedFile.length() > 0) {
+            Game::loadLevel(selectedFile);
+            delete dialog;
+            dialog = nullptr;
         }
     }
+    ImGui::SameLine();
+    if (ImGui::Button("Toggle Console")) {
+        Console::setVisible(!Console::isVisible());
+    }
+    ImGui::SameLine();
+    if (ImGui::Button("Toggle Debug Window")) {
+        UI::setVisible(!UI::isVisible());
+    }
+
+    ImGui::Separator();
+
+    // Set up columns for game list
+    ImGui::Columns(3, "menuscripts");
+    static bool offsets = false;
+    if (!offsets) {
+        ImGui::SetColumnOffset(1, 140.0f);
+        ImGui::SetColumnOffset(2, size.x - 200.0f);
+        offsets = true;
+    }
+
+    // List found games
+    for (int i = 0; i < scripts.size(); i++) {
+        ImGui::PushID(i);
+        Script& s = scripts.at(i);
+
+        if (images.at(i) >= 0) {
+            auto bm = TextureManager::getBufferManager(images.at(i), TextureStorage::SYSTEM);
+            ImGui::Image(bm, ImVec2(128, 128));
+        }
+        ImGui::NextColumn();
+
+        ImGui::TextWrapped("Language: %s", s.getLanguage().c_str());
+        ImGui::TextWrapped("Description: %s", s.getDescription().c_str());
+
+        if (ImGui::TreeNode("", "%d Levels", s.levelCount())) {
+            for (int l = 0; l < s.levelCount(); l++) {
+                ImGui::PushID(l);
+                ImGui::Bullet();
+                ImGui::TextWrapped("%s (%s)", s.getLevelName(l).c_str(), s.getLevelFilename(l).c_str());
+                ImGui::SameLine();
+                if (ImGui::Button("Play level")) {
+                    std::vector<File> levelFiles;
+                    paths.at(i).findRecursiveFilesEndingWith(levelFiles,
+                            getLastPathElement(s.getLevelFilename(l)));
+                    if (levelFiles.size() == 0) {
+                        Log::get(LOG_ERROR) << "Could not find level \""
+                                            << getLastPathElement(s.getLevelFilename(l))
+                                            << "\"!" << Log::endl;
+                    } else {
+                        Game::loadLevel(levelFiles.at(0).getPath());
+                    }
+                }
+                ImGui::PopID();
+            }
+            ImGui::TreePop();
+        }
+
+        if (ImGui::TreeNode("", "%d Cut-Scenes", s.cutsceneCount())) {
+            for (int c = 0; c < s.cutsceneCount(); c++) {
+                ImGui::Bullet();
+                ImGui::TextWrapped("%s", s.getCutsceneFilename(c).c_str());
+            }
+            ImGui::TreePop();
+        }
+
+        if (ImGui::TreeNode("", "%d Videos", s.videoCount())) {
+            for (int v = 0; v < s.videoCount(); v++) {
+                ImGui::Bullet();
+                ImGui::TextWrapped("%s", s.getVideoFilename(v).c_str());
+            }
+            ImGui::TreePop();
+        }
+
+        if (ImGui::TreeNode("", "%d Pictures", s.pictureCount())) {
+            for (int p = 0; p < s.pictureCount(); p++) {
+                ImGui::Bullet();
+                ImGui::TextWrapped("%s", s.getPictureFilename(p).c_str());
+            }
+            ImGui::TreePop();
+        }
+
+        if (ImGui::TreeNode("", "%d Titles", s.titleCount())) {
+            for (int t = 0; t < s.titleCount(); t++) {
+                ImGui::Bullet();
+                ImGui::TextWrapped("%s", s.getTitleFilename(t).c_str());
+            }
+            ImGui::TreePop();
+        }
+
+        ImGui::NextColumn();
+        ImGui::TextWrapped("Real Gameplay not yet implemented!");
+        ImGui::NextColumn();
+
+        if (i < (scripts.size() - 1))
+            ImGui::Separator();
+        ImGui::PopID();
+    }
+    ImGui::Columns(1);
+
+    if (scripts.size() == 0) {
+        ImGui::TextWrapped("OpenRaiders built-in TombRaider-Script detection mechanism could not find any suitable TR2/TR3 script files (called \"TOMBPC.DAT\") in your PAK folder \"%s\"! Use the \"Select Level File\" Button to load a single level without starting the entire game.",
+                           RunTime::getPakDir().c_str());
+    }
+
+    ImGui::End();
 }